+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix dev server proxy for default plugins repo to pass through module-scoped plugin urls
+
 ## [0.6.3] - 2022-09-04
 
 ### Added